Carry lsad as int64 to avoid truncating the scaled SAD limit
d->lsad was an int but was scaled by pixelMax and the block area
(computed in int64) before being stored back, truncating, while the
SearchMVs/DoSearchMVs lsad parameter was also int even though the
LSAD member it feeds is already int64. With a large user lsad at high
bitdepth and large blocks the value wrapped to a negative/garbage SAD
threshold. Widen the Analyse field and the lsad parameter chain to
int64 to match LSAD and the sibling badSAD.

+++ src/Analyse.cpp
@@ -32,7 +32,7 @@ struct AnalyseData {
     int nSearchParam; 
     int nPelSearch; 
 
-    int lsad; 
+    int64_t lsad;
     int pnew;   
     int plevel;    
     bool global;    
@@ -251,7 +251,7 @@ static void VS_CC analyseCreate(const VSMap *in, VSMap *out, [[maybe_unused]] vo
             d->chroma = false;
 
         int pixelMax = (1 << d->vi->format.bitsPerSample) - 1;
-        d->lsad = (int)((double)d->lsad * pixelMax / 255.0 + 0.5);
+        d->lsad = (int64_t)((double)d->lsad * pixelMax / 255.0 + 0.5);
         d->badSAD = (int)((double)d->badSAD * pixelMax / 255.0 + 0.5);
         d->nLambda = (int)((double)d->nLambda * pixelMax / 255.0 + 0.5);
